# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type VehicleType = (typeof VehicleType)[keyof typeof VehicleType];

export const VehicleType = {
  car: "car",
  bike: "bike",
  auto: "auto",
} as const;

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  confirmed: "confirmed",
  active: "active",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface ParkingLocation {
  id: number;
  name: string;
  city: string;
  area: string;
  address: string;
  landmark: string;
  operatorName: string;
  totalSlots: number;
  availableSlots: number;
  hourlyRate: number;
  vehicleType: VehicleType;
  openTime: string;
  closeTime: string;
  securityAvailable: boolean;
  evCharging: boolean;
  createdAt: string;
}

export interface ParkingLocationInput {
  name: string;
  city: string;
  area: string;
  address: string;
  landmark: string;
  operatorName: string;
  /** @minimum 1 */
  totalSlots: number;
  /** @minimum 0 */
  availableSlots: number;
  /** @minimum 0 */
  hourlyRate: number;
  vehicleType: VehicleType;
  openTime: string;
  closeTime: string;
  securityAvailable: boolean;
  evCharging: boolean;
}

export interface ParkingLocationUpdate {
  name?: string;
  city?: string;
  area?: string;
  address?: string;
  landmark?: string;
  operatorName?: string;
  /** @minimum 1 */
  totalSlots?: number;
  /** @minimum 0 */
  availableSlots?: number;
  /** @minimum 0 */
  hourlyRate?: number;
  vehicleType?: VehicleType;
  openTime?: string;
  closeTime?: string;
  securityAvailable?: boolean;
  evCharging?: boolean;
}

export type BookingPaymentMode =
  (typeof BookingPaymentMode)[keyof typeof BookingPaymentMode];

export const BookingPaymentMode = {
  cash: "cash",
  upi: "upi",
} as const;

export interface Booking {
  id: number;
  locationId: number;
  locationName: string;
  city: string;
  area: string;
  userName: string;
  phone: string;
  vehicleNumber: string;
  vehicleType: VehicleType;
  startTime: string;
  durationHours: number;
  amount: number;
  paymentMode: BookingPaymentMode;
  status: BookingStatus;
  createdAt: string;
}

export type BookingInputPaymentMode =
  (typeof BookingInputPaymentMode)[keyof typeof BookingInputPaymentMode];

export const BookingInputPaymentMode = {
  cash: "cash",
  upi: "upi",
} as const;

export interface BookingInput {
  locationId: number;
  userName: string;
  phone: string;
  vehicleNumber: string;
  vehicleType: VehicleType;
  startTime: string;
  /** @minimum 1 */
  durationHours: number;
  paymentMode: BookingInputPaymentMode;
}

export interface BookingUpdate {
  status?: BookingStatus;
}

export type ParkingSummaryCityBreakdownItem = {
  city: string;
  locations: number;
  availableSlots: number;
};

export interface ParkingSummary {
  totalLocations: number;
  totalSlots: number;
  availableSlots: number;
  activeBookings: number;
  todayRevenue: number;
  cityBreakdown: ParkingSummaryCityBreakdownItem[];
  recentBookings: Booking[];
}

export type ListParkingLocationsParams = {
  city?: string;
  vehicleType?: ListParkingLocationsVehicleType;
};

export type ListParkingLocationsVehicleType =
  (typeof ListParkingLocationsVehicleType)[keyof typeof ListParkingLocationsVehicleType];

export const ListParkingLocationsVehicleType = {
  car: "car",
  bike: "bike",
  auto: "auto",
} as const;

export type ListBookingsParams = {
  phone?: string;
  status?: ListBookingsStatus;
};

export type ListBookingsStatus =
  (typeof ListBookingsStatus)[keyof typeof ListBookingsStatus];

export const ListBookingsStatus = {
  confirmed: "confirmed",
  active: "active",
  completed: "completed",
  cancelled: "cancelled",
} as const;