# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking Management API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingPaymentMode } from "./bookingPaymentMode";
import type { BookingStatus } from "./bookingStatus";
import type { VehicleType } from "./vehicleType";

export interface Booking {
  id: number;
  locationId: number;
  locationName: string;
  city: string;
  area: string;
  userName: string;
  phone: string;
  vehicleNumber: string;
  vehicleType: VehicleType;
  startTime: string;
  durationHours: number;
  amount: number;
  paymentMode: BookingPaymentMode;
  status: BookingStatus;
  createdAt: string;
}