# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking Management API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingInputPaymentMode } from "./bookingInputPaymentMode";
import type { VehicleType } from "./vehicleType";

export interface BookingInput {
  locationId: number;
  userName: string;
  phone: string;
  vehicleNumber: string;
  vehicleType: VehicleType;
  startTime: string;
  /** @minimum 1 */
  durationHours: number;
  paymentMode: BookingInputPaymentMode;
}