# lib/api-zod/src/generated/types/bookingStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking Management API
 * OpenAPI spec version: 0.1.0
 */

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  confirmed: "confirmed",
  active: "active",
  completed: "completed",
  cancelled: "cancelled",
} as const;