# lib/api-zod/src/generated/types/bookingUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking Management API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";

export interface BookingUpdate {
  status?: BookingStatus;
}