# lib/api-zod/src/generated/types/parkingLocation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking Management API
 * OpenAPI spec version: 0.1.0
 */
import type { VehicleType } from "./vehicleType";

export interface ParkingLocation {
  id: number;
  name: string;
  city: string;
  area: string;
  address: string;
  landmark: string;
  operatorName: string;
  totalSlots: number;
  availableSlots: number;
  hourlyRate: number;
  vehicleType: VehicleType;
  openTime: string;
  closeTime: string;
  securityAvailable: boolean;
  evCharging: boolean;
  createdAt: string;
}