# lib/api-zod/src/generated/types/parkingLocationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking Management API
 * OpenAPI spec version: 0.1.0
 */
import type { VehicleType } from "./vehicleType";

export interface ParkingLocationInput {
  name: string;
  city: string;
  area: string;
  address: string;
  landmark: string;
  operatorName: string;
  /** @minimum 1 */
  totalSlots: number;
  /** @minimum 0 */
  availableSlots: number;
  /** @minimum 0 */
  hourlyRate: number;
  vehicleType: VehicleType;
  openTime: string;
  closeTime: string;
  securityAvailable: boolean;
  evCharging: boolean;
}