# lib/api-zod/src/generated/types/parkingSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Booking } from "./booking";
import type { ParkingSummaryCityBreakdownItem } from "./parkingSummaryCityBreakdownItem";

export interface ParkingSummary {
  totalLocations: number;
  totalSlots: number;
  availableSlots: number;
  activeBookings: number;
  todayRevenue: number;
  cityBreakdown: ParkingSummaryCityBreakdownItem[];
  recentBookings: Booking[];
}