# lib/api-zod/src/generated/types/vehicleType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking Management API
 * OpenAPI spec version: 0.1.0
 */

export type VehicleType = (typeof VehicleType)[keyof typeof VehicleType];

export const VehicleType = {
  car: "car",
  bike: "bike",
  auto: "auto",
} as const;